+++ packages/apollo-engine-reporting-protobuf/.gitignore
@@ -1,0 +1,2 @@
+protobuf.d.ts
+protobuf.js
