+++ packages/apollo-engine-reporting-protobuf/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../../tsconfig.base",
+  "compilerOptions": {
+    "rootDir": "./src",
+    "outDir": "./dist"
+  },
+  "include": ["src/**/*"],
+  "exclude": ["src/protobuf.d.ts", "src/protobuf.js"]
+}
